local-agent-test/requirements.txt
strands-agents[openai]>=1.6.0
strands-agents-tools
kubernetes>=28.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
requests>=2.31.0
